requirements.txt
paddlepaddle-gpu==1.6.2
numpy
scipy